requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.9.9,<3.0
dj_database_url>=2.1.0,<2.2
django-cors-headers==3.7.0
PyJWT==2.0.1
Faker==6.6.1
